pyproject.toml
[project]
name = "comfyui-mickmumpitz-nodes"
description = "A collection of custom nodes for ComfyUI by MickMumpitz"
version = "1.0.0"
license = { text = "MIT" }

authors = [
    { name = "MickMumpitz" }
]

requires-python = ">=3.8"

dependencies = [
    "torch",
    "numpy",
    "Pillow",
]

[project.urls]
Repository = "https://github.com/yourusername/comfyui-mickmumpitz-nodes"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
